requirements.txt
streamlit==1.46.0
pandas==2.2.2
numpy==2.0.2
nltk==3.9.1
tensorflow==2.18.0
sklearn-pandas==2.2.0
sklearn-compat==0.1.3
pyngrok==7.2.11